tavily_data_pipline/requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pydantic>=2.0.0

# Tavily API
tavily-python>=0.3.0

# OpenAI (for agent)
openai>=1.0.0

# MongoDB
pymongo>=4.6.0

# AWS
boto3>=1.34.0

# Snowflake
snowflake-connector-python>=3.7.0

# Streamlit
streamlit>=1.28.0
altair>=5.0.0

# Utilities
python-dateutil>=2.8.2
